createrepo: add support to remove old files
repomanage is now used to remove old binaries from the repository.
The default is to keep only the newest binary. This behavior can be modified
by setting JPB_REPO_KEEP to the amount of binaries to keep. If JPB_REPO_KEEP is 0
no cleanup is done at all.
The debian packages was also updated to install repomanage